chore: rename s3.ts to S3Client.ts
