+++ packages/backend/src/services/SemanticLayerService/SemanticLayerService.ts
@@ -17,7 +17,7 @@ import {
     type AbilityAction,
 } from '@lightdash/common';
 import { LightdashAnalytics } from '../../analytics/LightdashAnalytics';
-import { S3Client } from '../../clients/Aws/s3';
+import { S3Client } from '../../clients/Aws/S3Client';
 import CubeClient from '../../clients/cube/CubeClient';
 import DbtCloudGraphqlClient from '../../clients/dbtCloud/DbtCloudGraphqlClient';
 import { LightdashConfig } from '../../config/parseConfig';
